+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 * [UICONSET-127](https://issues.folio.org/browse/UICONSET-127) Implement `<ConsortiaControlledVolabulary>` component for consortium manager.
 * [UICONSET-59](https://issues.folio.org/browse/UICONSET-59) Consortia manager - Save/edit permissions set and assign users.
 * [UICONSET-111](https://issues.folio.org/browse/UICONSET-111) Apply publish coordinator to consortium member setting requests (without shared settings).
+* [UICONSET-113](https://issues.folio.org/browse/UICONSET-113) Allow a user to share settings in the `<ConsortiaControlledVocabulary>`.